+++ app/encryption_test.go
@@ -1,40 +1,32 @@
 package app
 
 import (
-	"fmt"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 )
 
-var original = "I love }apples{"
-var samplePasswords = []struct {
-	Description string
-	Password    string
-}{
-	{"all_numeric", `01234`},
-	{"alpha_lower", `abcdefg`},
-	{"alpha_upper", `ABCDEFG`},
-	{"empty_string", ``},
-	{"single_letter", `a`},
-	{"single_number", `5`},
-	{"special_chars", `~!@#$%^&*()_+-={}[]\\|;':\",./<>?`},
+func TestDefaultEncryption(t *testing.T) {
+	password := "my secret password"
+	content := "Some beer pls"
+	note := Note{Content: content, Password: password}
+	cipherText, cipherType, err := Encrypt(note)
+	assert.Nil(t, err, "Should not be an error calling encrypt")
+	assert.Equal(t, cipherType, "AES-GCM", "Default encryption type is wrong")
+	note.CipherType = cipherType
+	note.Content = cipherText
+	decrypted, err := Decrypt(note, password)
+	assert.Nil(t, err, "Should not be an error calling decrypt")
+	assert.Equal(t, decrypted, content, "Decryption should return the original")
 }
 
-func TestDecryptionReturnsOriginal(t *testing.T) {
-	for _, tt := range samplePasswords {
-		encrypted, err := Encrypt(original, tt.Password)
-		assert.Nil(t, err, "Should not be an error calling encrypt")
-		decrypted, err := Decrypt(encrypted, tt.Password)
-		assert.Nil(t, err, "Should not be an error calling decrypt")
-		msg := fmt.Sprintf("Decryption should return the original (%s)", tt.Description)
-		assert.Equal(t, decrypted, original, msg)
-	}
+func TestDecryptionOfDeprecatedCBC(t *testing.T) {
 	password := "my secret password"
-	encrypted, err := Encrypt(original, password)
+	content := "Some beer pls"
+	encrypted, err := CBCEncrypt(content, password)
 	assert.Nil(t, err, "Should not be an error calling encrypt")
-	decrypted, err := Decrypt(encrypted, password)
+	note := Note{Content: encrypted, Encrypted: true}
+	decrypted, err := Decrypt(note, password)
 	assert.Nil(t, err, "Should not be an error calling decrypt")
-	assert.Equal(t, decrypted, original, "Decryption should return the original")
-
+	assert.Equal(t, decrypted, content, "Decryption should return the original")
 }